move code to a routerfs package
