+++ src/test/java/io/lakefs/routerfs/PathMapperTest.java
@@ -1,5 +1,6 @@
-package io.lakefs;
+package io.lakefs.routerfs;
 
+import io.lakefs.routerfs.PathMapper;
 import org.apache.hadoop.conf.Configuration;
 import org.junit.Assert;
 import org.junit.Rule;
